move ChildObserver into observers.py
