+++ src/vivarium_gates_nutrition_optimization/components/__init__.py
@@ -1,4 +1,3 @@
-from .children import BirthObserver
 from .hemoglobin import Anemia, Hemoglobin
 from .intervention import MaternalInterventions
 from .maternal_bmi import MaternalBMIExposure
@@ -7,6 +6,7 @@ from .morbidity import BackgroundMorbidity
 from .mortality import MaternalMortality
 from .observers import (
     AnemiaObserver,
+    BirthObserver,
     DisabilityObserver,
     MaternalBMIObserver,
     MaternalInterventionObserver,
